Add files via upload
So. I let this render for 24 hours on gpu compute with the samples at 300 and it did not render fully. I got what I was looking for though. I'm sure this MacBook Pro will pay off, although I don't know if it's best for blender. I also don't want this to be a vendetta against anything other than malignant forces.